Refatora componentes reutilizáveis
