+++ globoplay-iOS/globoplay-iOS/Scenes/MediaDetails/MediaDetailsView.swift
@@ -56,6 +56,8 @@ struct MediaDetailsView: View {
                     .padding(.horizontal, 8)
                 
                 myListButton
+                
+                
             }
         }
     }
